backpython/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.35.71
